refactor(types): Fix types
