--- apps/web/types/blog.types.ts
@@ -1,8 +1,0 @@
-export interface BlogMetadata {
-  title: string;
-  description?: string;
-  date?: string;
-  tags?: string[];
-  image?: string;
-  wordCount?: number; // New property for word count
-}
