--- apps/web/types/portfolio.types.ts
@@ -1,7 +1,0 @@
-export interface PortfolioMetadata {
-  title: string;
-  description?: string;
-  date?: string;
-  tags?: string[];
-  image?: string;
-}
